requirements.txt
pillow==12.0.0
python-musicpd==0.9.2
RPi.GPIO==0.7.1
mediafile==0.13.0
numpy==2.4.0
st7789==1.0.1
PyYAML==6.0.3
spidev==3.8
gpiod==2.4.0
gpiodevice==0.0.5
